% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as.list.robotstxt_text}
\alias{as.list.robotstxt_text}
\title{Convert robotstxt_text to list}
\usage{
\method{as.list}{robotstxt_text}(x, ...)
}
\arguments{
\item{x}{class robotstxt_text object to be transformed into list}

\item{...}{further arguments (inherited from \code{base::as.list()})}
}
\description{
Convert robotstxt_text to list
}
