% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{get_comments}
\alias{get_comments}
\title{function for extrcting comments from robots.txt}
\usage{
get_comments(txt)
}
\arguments{
\item{txt}{content of the robots.txt file}
}
\description{
function for extrcting comments from robots.txt
}

