% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{get_fields}
\alias{get_fields}
\title{function for extracting robotstxt fields}
\usage{
get_fields(txt, type = "all", regex = NULL, invert = FALSE)
}
\arguments{
\item{txt}{content of the robots.txt file}
}
\description{
function for extracting robotstxt fields
}

