% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{get_permissions}
\alias{get_permissions}
\title{function for extracting permissions from robots.txt}
\usage{
get_permissions(txt)
}
\arguments{
\item{txt}{content of the robots.txt file}
}
\description{
function for extracting permissions from robots.txt
}

