% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{get_useragent}
\alias{get_useragent}
\title{function for extracting HTTP useragents from robots.txt}
\usage{
get_useragent(txt)
}
\arguments{
\item{txt}{content of the robots.txt file}
}
\description{
function for extracting HTTP useragents from robots.txt
}

