% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_domain.R
\name{guess_domain}
\alias{guess_domain}
\title{function guessing domain from path}
\usage{
guess_domain(x)
}
\arguments{
\item{x}{path aka URL from which to infer domain}
}
\description{
function guessing domain from path
}
