% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellno.R
\docType{data}
\name{HELLNO}
\alias{HELLNO}
\title{logical constant for FALSE}
\format{An object of class \code{logical} of length 1.}
\usage{
HELLNO
}
\description{
logical constant for FALSE
}
\keyword{datasets}

