% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{named_list}
\alias{named_list}
\title{make automatically named list}
\usage{
named_list(...)
}
\arguments{
\item{...}{things to be put in list}
}
\description{
make automatically named list
}
\keyword{internal}
