% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_to_default.R
\name{null_to_defeault}
\alias{null_to_defeault}
\title{null_to_defeault}
\usage{
null_to_defeault(x, d)
}
\arguments{
\item{x}{value to check and return}

\item{d}{value to return in case x is NULL}
}
\description{
null_to_defeault
}
