% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{path_allowed}
\alias{path_allowed}
\title{check if a bot has permissions to access page}
\usage{
path_allowed(permissions, path = "/", bot = "*")
}
\arguments{
\item{permissions}{data.frame generated by \code{rt_get_permissions()} with
three columns: 'useragent' (name of the bot), 'permission' (permission
granted: Disallow/Allow) and 'value' (path for which permission applies)}

\item{path}{path for which to check bot's permission, defaults to "/"}

\item{bot}{name of the bot, defaults to "*"}
}
\value{
TRUE permission granted; FALSE permission denied; NA path_allowed()
cannot determine if path is allowed or not allowed, will result in a warning
as well
}
\description{
check if a bot has permissions to access page
}
\seealso{
\link{paths_allowed}
}
\keyword{internal}

