% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_allowed.R
\name{paths_allowed_worker_robotstxt}
\alias{paths_allowed_worker_robotstxt}
\title{paths_allowed_worker for robotstxt flavor}
\usage{
paths_allowed_worker_robotstxt(user_agent, domain, bot, paths)
}
\description{
paths_allowed_worker for robotstxt flavor
}
