% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_allowed.R
\name{paths_allowed_worker_spiderbar}
\alias{paths_allowed_worker_spiderbar}
\title{paths_allowed_worker spiderbar flavor}
\usage{
paths_allowed_worker_spiderbar(user_agent, domain, bot)
}
\arguments{
\item{bot}{}
}
\description{
paths_allowed_worker spiderbar flavor
}
