% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_robotstxt.R
\name{print.robotstxt}
\alias{print.robotstxt}
\title{Print robotstxt}
\usage{
\method{print}{robotstxt}(x, ...)
}
\arguments{
\item{x}{robotstxt instance to be printed}

\item{...}{goes down the sink}
}
\description{
Print robotstxt
}
