% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotstxt.R
\name{print.robotstxt_text}
\alias{print.robotstxt_text}
\title{printing robotstxt_text}
\usage{
\method{print}{robotstxt_text}(x, ...)
}
\arguments{
\item{x}{character vector aka robotstxt$text to be printed}

\item{...}{goes down the sink}
}
\description{
printing robotstxt_text
}

