% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellno.R
\name{rbind}
\alias{rbind}
\title{altenative rbind implementation}
\usage{
rbind(..., deparse.level = 1)
}
\arguments{
\item{...}{see \code{\link[base]{rbind.data.frame}}}

\item{deparse.level}{see \code{\link[base]{rbind.data.frame}}}
}
\description{
altenative rbind implementation
}
\seealso{
\code{\link[base]{rbind.data.frame}}
}

