% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_domain.R
\name{remove_domain}
\alias{remove_domain}
\title{Remove domain from path}
\usage{
remove_domain(x)
}
\arguments{
\item{x}{path aka URL from which to first infer domain and then remove it}
}
\description{
Remove domain from path
}
