% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_cache.R
\docType{data}
\name{rt_cache}
\alias{rt_cache}
\title{Get the robotstxt cache}
\format{
An object of class \code{environment} of length 0.
}
\usage{
rt_cache
}
\description{
Get the robotstxt cache
}
\keyword{datasets}
