% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{rt_get_fields_worker}
\alias{rt_get_fields_worker}
\title{extracting robotstxt fields}
\usage{
rt_get_fields_worker(txt, type = "all", regex = NULL, invert = FALSE)
}
\arguments{
\item{txt}{content of the robots.txt file}

\item{type}{name or names of the fields to be returned, defaults to all
fields}

\item{regex}{subsetting field names via regular expressions}

\item{invert}{field selection}
}
\description{
extracting robotstxt fields
}
\keyword{internal}
