% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{rt_get_permissions}
\alias{rt_get_permissions}
\title{extracting permissions from robots.txt}
\usage{
rt_get_permissions(txt)
}
\arguments{
\item{txt}{content of the robots.txt file}
}
\description{
extracting permissions from robots.txt
}

