% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{rt_get_robotstxt}
\alias{rt_get_robotstxt}
\title{downloading robots.txt file}
\usage{
rt_get_robotstxt(domain)
}
\arguments{
\item{domain}{domain from which to download robots.txt file}
}
\description{
downloading robots.txt file
}

