% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rt_get_rtxt}
\alias{rt_get_rtxt}
\title{load robots.txt files saved along with the package}
\usage{
rt_get_rtxt(name = sample(rt_list_rtxt(), 1))
}
\arguments{
\item{name}{name of the robots.txt files, defaults to a random drawn file ;-)}
}
\description{
load robots.txt files saved along with the package:
these functions ar very handy for testing (not used otherwise)
}
\keyword{internal}
