% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rt_list_rtxt}
\alias{rt_list_rtxt}
\title{list robots.txt files saved along with the package}
\usage{
rt_list_rtxt()
}
\description{
list robots.txt files saved along with the package:
these functions ar very handy for testing (not used otherwise)
}
\keyword{internal}

