% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_request_handler.R
\name{rt_request_handler}
\alias{rt_request_handler}
\title{rt_request_handler}
\usage{
rt_request_handler(request,
  rt_event_handler = robotstxt::default_event_handler, warn = TRUE,
  encoding = "UTF-8")
}
\arguments{
\item{request}{result of an HTTP request (e.g. httr::GET())}

\item{warn}{suppress warnings}

\item{encoding}{The text encoding to assume if no encoding is provided in the
headers of the response}
}
\value{
a list with three items following the following schema: \cr \code{
  list( rtxt = "", problems = list( "redirect" = list( status_code = 301 ),
  "domain" = list(from_url = "...", to_url = "...") ) ) }
}
\description{
A helper function for get_robotstxt() that will extract the robots.txt file
from the HTTP request result object. furthermore it will inform
get_robotstxt() if the request should be cached and which problems occured.
}
