% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{sanatize_path}
\alias{sanatize_path}
\title{function for making paths uniform}
\usage{
sanatize_path(path)
}
\arguments{
\item{path}{path to be sanatized}
}
\value{
sanatized path
}
\description{
function for making paths uniform
}

