% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{sanitize_permission_values}
\alias{sanitize_permission_values}
\title{transfroming permissions into regular expressions}
\usage{
sanitize_permission_values(permission_value)
}
\arguments{
\item{permissions}{the permissions to be transformed}
}
\description{
transfroming permissions into regular expressions
}

