% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_permissions.R
\name{sanitize_permissions}
\alias{sanitize_permissions}
\title{transforming permissions into regular expressions (whole permission)}
\usage{
sanitize_permissions(permissions)
}
\arguments{
\item{permissions}{the permissions to be transformed}
}
\description{
transforming permissions into regular expressions (whole permission)
}
\keyword{internal}
