% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdomain_changed_to_www.R
\name{subdomain_changed_to_www}
\alias{subdomain_changed_to_www}
\title{subdomain_changed_to_www}
\usage{
subdomain_changed_to_www(response)
}
\arguments{
\item{response}{an httr response object, e.g. from a call to httr::GET()}
}
\value{
logical of length 1 indicating whether or not any domain change
    happened during the HTTP request
}
\description{
subdomain_changed_to_www
}
