% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_request_handler_defaults.R
\docType{data}
\name{default_event_handler}
\alias{default_event_handler}
\title{default HTTP request event handler}
\format{An object of class \code{list} of length 7.}
\usage{
default_event_handler
}
\description{
default HTTP request event handler
}
\keyword{datasets}
