% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{get_fields}
\alias{get_fields}
\title{function for extracting robotstxt fields}
\usage{
get_fields(txt, type = "all", regex = NULL, invert = FALSE)
}
\arguments{
\item{txt}{content of the robots.txt file}

\item{type}{name or names of the fields to be returned, defaults to all
fields}

\item{regex}{subsetting field names via regular expressions}

\item{invert}{field selection}
}
\description{
function for extracting robotstxt fields
}

