% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{get_robotstxt}
\alias{get_robotstxt}
\title{downloading robots.txt file}
\usage{
get_robotstxt(domain, warn = TRUE)
}
\arguments{
\item{domain}{domain from which to download robots.txt file}

\item{warn}{warn about being unable to download domain/robots.txt because of
HTTP resposne status 404. If this happens,}
}
\description{
downloading robots.txt file
}

