% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_subdomain_changed.R
\name{http_subdomain_changed}
\alias{http_subdomain_changed}
\title{http_subdomain_changed}
\usage{
http_subdomain_changed(response)
}
\arguments{
\item{response}{an httr response object, e.g. from a call to httr::GET()}
}
\value{
logical of length 1 indicating whether or not any domain change
    happened during the HTTP request
}
\description{
http_subdomain_changed
}
