% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_was_redirected.R
\name{http_was_redirected}
\alias{http_was_redirected}
\title{http_was_redirected}
\usage{
http_was_redirected(response)
}
\arguments{
\item{response}{an httr response object, e.g. from a call to httr::GET()}
}
\value{
logical of length 1 indicating whether or not any redirect happened
  during the HTTP request
}
\description{
http_was_redirected
}
