% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_suspect_robotstxt.R
\name{is_suspect_robotstxt}
\alias{is_suspect_robotstxt}
\title{is_suspect_robotstxt}
\usage{
is_suspect_robotstxt(text)
}
\arguments{
\item{text}{content of a robots.txt file provides as character vector}
}
\description{
function that checks if file is valid / parsable robots.txt file
}
