% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_to_default.R
\name{null_to_default}
\alias{null_to_default}
\title{Return default value if NULL}
\usage{
null_to_default(x, d)
}
\arguments{
\item{x}{value to check and return}

\item{d}{value to return in case x is NULL}
}
\description{
Return default value if NULL
}
