% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{parse_robotstxt}
\alias{parse_robotstxt}
\title{function parsing robots.txt}
\usage{
parse_robotstxt(txt)
}
\arguments{
\item{txt}{content of the robots.txt file}
}
\value{
a named list with useragents, comments, permissions, sitemap
}
\description{
function parsing robots.txt
}

