% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{paths_allowed}
\alias{paths_allowed}
\title{check if a bot has permissions to access page}
\usage{
paths_allowed(permissions, paths = "/", bot = "*")
}
\arguments{
\item{permissions}{data.frame generated by \code{rt_get_permissions()} with
three columns: 'useragent' (name of the bot), 'permission' (permission
granted: Disallow/Allow) and 'value' (path for which permission applies)}

\item{paths}{paths for which to check bot's permission, defaults to "/"}

\item{bot}{name of the bot, defaults to "*"}
}
\description{
wrapper to \code{\link{path_allowed}}
}
\seealso{
\link{path_allowed}
}
\keyword{internal}

