% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{paths_allowed}
\alias{paths_allowed}
\title{check if a bot has permissions to access page(s)}
\usage{
paths_allowed(paths = "/", domain = "auto", bot = "*")
}
\arguments{
\item{paths}{paths for which to check bot's permission, defaults to "/"}

\item{domain}{Domain for which paths should be checked. Defaults to "auto".
If set to "auto" function will try to guess the domain by parsing the paths
argument. Note however, that these are educated guesses which might utterly
fail. To be on the save side, provide appropriate domains manually.}

\item{bot}{name of the bot, defaults to "*"}
}
\description{
wrapper to \code{\link{path_allowed}}
}
\seealso{
\link{path_allowed}
}

