% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_allowed_worker_robotstxt.R
\name{paths_allowed_worker_robotstxt}
\alias{paths_allowed_worker_robotstxt}
\title{paths_allowed_worker for robotstxt flavor}
\usage{
paths_allowed_worker_robotstxt(domain, bot, paths, permissions_list)
}
\arguments{
\item{domain}{Domain for which paths should be checked. Defaults to "auto".
If set to "auto" function will try to guess the domain by parsing the paths
argument. Note however, that these are educated guesses which might utterly
fail. To be on the save side, provide appropriate domains manually.}

\item{bot}{name of the bot, defaults to "*"}

\item{paths}{paths for which to check bot's permission, defaults to "/"}

\item{permissions_list}{list of permissions from robotstxt objects}
}
\description{
paths_allowed_worker for robotstxt flavor
}
