% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_allowed.R
\name{paths_allowed_worker_robotstxt}
\alias{paths_allowed_worker_robotstxt}
\title{paths_allowed_worker for robotstxt flavor}
\usage{
paths_allowed_worker_robotstxt(user_agent, domain, bot, paths)
}
\arguments{
\item{user_agent}{HTTP user-agent string to be used to retrieve robots.txt file
from domain}

\item{domain}{Domain for which paths should be checked. Defaults to "auto".
If set to "auto" function will try to guess the domain by parsing the paths
argument. Note however, that these are educated guesses which might utterly
fail. To be on the save side, provide appropriate domains manually.}

\item{bot}{name of the bot, defaults to "*"}

\item{paths}{paths for which to check bot's permission, defaults to "/"}
}
\description{
paths_allowed_worker for robotstxt flavor
}
