% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotstxt.R
\docType{data}
\name{robotstxt}
\alias{robotstxt}
\title{An object representation of robotstxt}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rt <- robotstxt$new(domain="google.com")
rt$bots
rt$permissions
}
\description{
An object representation of robotstxt
}
\section{Fields}{

\describe{
\item{\code{text}}{text of robots.txt either supplied by user or downloaded from
domain}

\item{\code{bots}}{vector of bot names mentionend in robots.txt}

\item{\code{permissions}}{Data.frame with all permissions (Allow/Disallow) the bot
name they apply to and the path ().}

\item{\code{domain}}{Domain for which the robots.txt file is valid.}

\item{\code{sitemap}}{Robots.txt files might contain sitemap entries -- these are
accessible as data.frame via this field.}

\item{\code{other}}{Other fields that might have bee written in the robots.txt file.}

\item{\code{initialize(domain,text)}}{Method called when initialising object via
\code{robotstxt$new()}. Needs either \code{text} or \code{domain} to be
present at initialization. If only domain is supplied -- should be seen as
default -- than the robots.txt file for that domain will be downloaded. If
\code{text} is supplied as well, nothing will be downloaded. If only
\code{text} is supplied than domain is set to '???'.}

\item{\code{check(path="/",bot="*")}}{Method for checking whether or not a certain path is
allowed to be accessed by certain bot.}
}}
\keyword{datasets}

