% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\docType{data}
\name{rt_cache}
\alias{rt_cache}
\title{get_robotstxt() cache}
\format{An object of class \code{environment} of length 0.}
\usage{
rt_cache
}
\description{
get_robotstxt() cache
}
\keyword{datasets}
