% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{rt_get_comments}
\alias{rt_get_comments}
\title{extrcting comments from robots.txt}
\usage{
rt_get_comments(txt)
}
\arguments{
\item{txt}{content of the robots.txt file}
}
\description{
extrcting comments from robots.txt
}
\keyword{internal}
