% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{rt_get_fields}
\alias{rt_get_fields}
\title{extracting permissions from robots.txt}
\usage{
rt_get_fields(txt, regex = "", invert = FALSE)
}
\arguments{
\item{txt}{content of the robots.txt file}

\item{regex}{regular expression specify field}

\item{invert}{invert selection made via regex?}
}
\description{
extracting permissions from robots.txt
}
\keyword{internal}
