% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{rt_get_other}
\alias{rt_get_other}
\title{get_fields() wrapper to extract !(sitemap,(dis)allow,user-agent)}
\usage{
rt_get_other(txt, regex = "sitemap|allow|user-agent", invert = TRUE)
}
\arguments{
\item{txt}{\code{\link{rt_get_fields}}}

\item{regex}{defaults to "sitemap|allow|user-agent" ; \code{\link{rt_get_fields}}}

\item{invert}{defaults to TRUE; \code{\link{rt_get_fields}}}
}
\description{
get_fields() wrapper to extract !(sitemap,(dis)allow,user-agent)
}

