% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_robotstxt.R
\name{rt_get_useragent}
\alias{rt_get_useragent}
\title{extracting HTTP useragents from robots.txt}
\usage{
rt_get_useragent(txt)
}
\arguments{
\item{txt}{content of the robots.txt file}
}
\description{
extracting HTTP useragents from robots.txt
}
\keyword{internal}
