% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_path.R
\name{sanitize_path}
\alias{sanitize_path}
\title{making paths uniform}
\usage{
sanitize_path(path)
}
\arguments{
\item{path}{path to be sanitized}
}
\value{
sanitized path
}
\description{
making paths uniform
}
\keyword{internal}
