% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{sanitize_permission_values}
\alias{sanitize_permission_values}
\title{transforming permissions into regular expressions (values)}
\usage{
sanitize_permission_values(permission_value)
}
\arguments{
\item{permission_value}{the value column of permissions (the paths)}
}
\description{
transforming permissions into regular expressions (values)
}
\keyword{internal}

