% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{sanitize_permissions}
\alias{sanitize_permissions}
\title{transfroming permissions into regular expressions}
\usage{
sanitize_permissions(permissions)
}
\arguments{
\item{permissions}{the permissions to be transformed}
}
\description{
transfroming permissions into regular expressions
}

